UIKit extension of PositionAttribute
